prompt_recommendation/requirements.txt
streamlit==1.32.0
supabase==2.3.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.0
scikit-learn==1.3.2
python-dotenv==1.0.0
huggingface_hub==0.25.0